+++ src/main/java/com/uspray/uspray/infrastructure/querydsl/pray/PrayRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.uspray.uspray.infrastructure;
+package com.uspray.uspray.infrastructure.querydsl.pray;
 
 import com.uspray.uspray.domain.Pray;
 import java.util.List;
